+++ engines/twine/renderer/renderer.h
@@ -185,16 +185,16 @@ private:
 
 	bool _isUsingOrthoProjection = false;
 
-	void renderPolygonsCopper(int vtop, int32 vsize, uint8 color) const;
-	void renderPolygonsBopper(int vtop, int32 vsize, uint8 color) const;
-	void renderPolygonsFlat(int vtop, int32 vsize, uint8 color) const;
-	void renderPolygonsTele(int vtop, int32 vsize, uint8 color) const;
-	void renderPolygonsTrans(int vtop, int32 vsize, uint8 color) const;
-	void renderPolygonsTrame(int vtop, int32 vsize, uint8 color) const;
+	void renderPolygonsCopper(int vtop, int32 vsize, uint16 color) const;
+	void renderPolygonsBopper(int vtop, int32 vsize, uint16 color) const;
+	void renderPolygonsFlat(int vtop, int32 vsize, uint16 color) const;
+	void renderPolygonsTele(int vtop, int32 vsize, uint16 color) const;
+	void renderPolygonsTrans(int vtop, int32 vsize, uint16 color) const;
+	void renderPolygonsTrame(int vtop, int32 vsize, uint16 color) const;
 	void renderPolygonsGouraud(int vtop, int32 vsize) const;
 	void renderPolygonsDither(int vtop, int32 vsize) const;
 	void renderPolygonsMarble(int vtop, int32 vsize, uint8 colorStart, uint8 colorEnd) const;
-	void renderPolygonsSimplified(int vtop, int32 vsize, uint8 color) const;
+	void renderPolygonsSimplified(int vtop, int32 vsize, uint16 color) const;
 	void computePolygons(int16 polyRenderType, const Vertex *vertices, int32 numVertices);
 
 	const RenderCommand *depthSortRenderCommands(int32 numOfPrimitives);
